requirements/base.txt
# =============================================================
# ImoOS — Base Requirements
# =============================================================

# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8

# Multi-Tenancy
django-tenants==3.5.0

# Authentication & Permissions
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
django-allauth==0.60.1

# Audit & History
django-simple-history==3.4.0

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Geo / PostGIS
# GDAL installed in Dockerfile to match system libgdal version (see docker/Dockerfile.dev)
Shapely==2.0.2
djangorestframework-gis==1.0

# API Documentation
drf-spectacular==0.27.1

# Async / Background Tasks
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Storage
django-storages==1.14.2
boto3==1.34.0
Pillow==10.2.0

# Security
django-ratelimit==4.1.0
django-cryptography==1.1
bandit==1.7.7
django-csp==3.8
django-permissions-policy==4.28.0

# i18n
django-modeltranslation==0.18.11

# PDF Generation
WeasyPrint==60.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0
celery[redis]==5.3.6

# Observability (Sprint 7)
sentry-sdk[django]==1.40.0
django-prometheus==2.3.1
python-json-logger==2.0.7
